robust links types
